database component example
