+++ examples/database_components/src/lib.rs
@@ -1,8 +1,6 @@
 //! Optimism-specific constants, types, and helpers.
 #![cfg_attr(not(test), warn(unused_crate_dependencies))]
 
-
-
 //! Database that is split on State and BlockHash traits.
 pub mod block_hash;
 pub mod state;
@@ -10,12 +8,11 @@ pub mod state;
 pub use block_hash::{BlockHash, BlockHashRef};
 pub use state::{State, StateRef};
 
-use crate::{
-    db::{Database, DatabaseRef},
-    Account, AccountInfo, Address, Bytecode, HashMap, B256, U256,
+use revm::{
+    database_interface::{Database, DatabaseCommit, DatabaseRef},
+    primitives::{Address, HashMap, B256, U256},
+    state::{Account, AccountInfo, Bytecode},
 };
-
-use super::DatabaseCommit;
 
 #[derive(Debug)]
 pub struct DatabaseComponents<S, BH> {
@@ -84,127 +81,5 @@ impl<S: StateRef, BH: BlockHashRef> DatabaseRef for DatabaseComponents<S, BH> {
 impl<S: DatabaseCommit, BH: BlockHashRef> DatabaseCommit for DatabaseComponents<S, BH> {
     fn commit(&mut self, changes: HashMap<Address, Account>) {
         self.state.commit(changes);
-    }
-}
-
-
-//! BlockHash database component from [`database::Database`]
-//! it is used inside [`database::DatabaseComponents`]
-
-use crate::B256;
-use auto_impl::auto_impl;
-use core::ops::Deref;
-use std::sync::Arc;
-
-#[auto_impl(&mut, Box)]
-pub trait BlockHash {
-    type Error;
-
-    /// Get block hash by block number
-    fn block_hash(&mut self, number: u64) -> Result<B256, Self::Error>;
-}
-
-#[auto_impl(&, &mut, Box, Rc, Arc)]
-pub trait BlockHashRef {
-    type Error;
-
-    /// Get block hash by block number
-    fn block_hash(&self, number: u64) -> Result<B256, Self::Error>;
-}
-
-impl<T> BlockHash for &T
-where
-    T: BlockHashRef,
-{
-    type Error = <T as BlockHashRef>::Error;
-
-    fn block_hash(&mut self, number: u64) -> Result<B256, Self::Error> {
-        BlockHashRef::block_hash(*self, number)
-    }
-}
-
-impl<T> BlockHash for Arc<T>
-where
-    T: BlockHashRef,
-{
-    type Error = <T as BlockHashRef>::Error;
-
-    fn block_hash(&mut self, number: u64) -> Result<B256, Self::Error> {
-        self.deref().block_hash(number)
-    }
-}
-
-
-//! State database component from [`crate::db::Database`]
-//! it is used inside [`crate::db::DatabaseComponents`]
-
-use crate::{AccountInfo, Address, Bytecode, B256, U256};
-use auto_impl::auto_impl;
-use core::ops::Deref;
-use std::sync::Arc;
-
-#[auto_impl(&mut, Box)]
-pub trait State {
-    type Error;
-
-    /// Get basic account information.
-    fn basic(&mut self, address: Address) -> Result<Option<AccountInfo>, Self::Error>;
-
-    /// Get account code by its hash
-    fn code_by_hash(&mut self, code_hash: B256) -> Result<Bytecode, Self::Error>;
-
-    /// Get storage value of address at index.
-    fn storage(&mut self, address: Address, index: U256) -> Result<U256, Self::Error>;
-}
-
-#[auto_impl(&, &mut, Box, Rc, Arc)]
-pub trait StateRef {
-    type Error;
-
-    /// Get basic account information.
-    fn basic(&self, address: Address) -> Result<Option<AccountInfo>, Self::Error>;
-
-    /// Get account code by its hash
-    fn code_by_hash(&self, code_hash: B256) -> Result<Bytecode, Self::Error>;
-
-    /// Get storage value of address at index.
-    fn storage(&self, address: Address, index: U256) -> Result<U256, Self::Error>;
-}
-
-impl<T> State for &T
-where
-    T: StateRef,
-{
-    type Error = <T as StateRef>::Error;
-
-    fn basic(&mut self, address: Address) -> Result<Option<AccountInfo>, Self::Error> {
-        StateRef::basic(*self, address)
-    }
-
-    fn code_by_hash(&mut self, code_hash: B256) -> Result<Bytecode, Self::Error> {
-        StateRef::code_by_hash(*self, code_hash)
-    }
-
-    fn storage(&mut self, address: Address, index: U256) -> Result<U256, Self::Error> {
-        StateRef::storage(*self, address, index)
-    }
-}
-
-impl<T> State for Arc<T>
-where
-    T: StateRef,
-{
-    type Error = <T as StateRef>::Error;
-
-    fn basic(&mut self, address: Address) -> Result<Option<AccountInfo>, Self::Error> {
-        self.deref().basic(address)
-    }
-
-    fn code_by_hash(&mut self, code_hash: B256) -> Result<Bytecode, Self::Error> {
-        self.deref().code_by_hash(code_hash)
-    }
-
-    fn storage(&mut self, address: Address, index: U256) -> Result<U256, Self::Error> {
-        self.deref().storage(address, index)
     }
 }
